+++ core/trino-main/src/main/java/io/trino/operator/scalar/timestamp/ExtractDayOfYear.java
@@ -30,14 +30,14 @@ public class ExtractDayOfYear
     private ExtractDayOfYear() {}
 
     @LiteralParameters("p")
-    @SqlType(StandardTypes.BIGINT)
+    @SqlType(StandardTypes.SMALLINT)
     public static long extract(@SqlType("timestamp(p)") long timestamp)
     {
         return ISOChronology.getInstanceUTC().dayOfYear().get(scaleEpochMicrosToMillis(timestamp));
     }
 
     @LiteralParameters("p")
-    @SqlType(StandardTypes.BIGINT)
+    @SqlType(StandardTypes.SMALLINT)
     public static long extract(@SqlType("timestamp(p)") LongTimestamp timestamp)
     {
         return extract(timestamp.getEpochMicros());
